+++ automl/google/cloud/automl_v1beta1/__init__.py
@@ -21,12 +21,16 @@ from google.cloud.automl_v1beta1 import types
 from google.cloud.automl_v1beta1.gapic import auto_ml_client
 from google.cloud.automl_v1beta1.gapic import enums
 from google.cloud.automl_v1beta1.gapic import prediction_service_client
-from google.cloud.automl_v1beta1.tables import gcs_client
 from google.cloud.automl_v1beta1.tables import tables_client
+from google.cloud.automl_v1beta1.tables import gcs_client
 
 
 class TablesClient(tables_client.TablesClient):
     __doc__ = tables_client.TablesClient.__doc__
+
+
+class GcsClient(gcs_client.GcsClient):
+    __doc__ = gcs_client.GcsClient.__doc__
 
 
 class AutoMlClient(auto_ml_client.AutoMlClient):
@@ -37,10 +41,6 @@ class AutoMlClient(auto_ml_client.AutoMlClient):
 class PredictionServiceClient(prediction_service_client.PredictionServiceClient):
     __doc__ = prediction_service_client.PredictionServiceClient.__doc__
     enums = enums
-
-
-class GcsClient(gcs_client.GcsClient):
-    __doc__ = gcs_client.GcsClient.__doc__
 
 
 __all__ = (