+++ automl/noxfile.py
@@ -69,7 +69,6 @@ def default(session):
     session.install("mock", "pytest", "pytest-cov")
     for local_dep in LOCAL_DEPS:
         session.install("-e", local_dep)
-    session.install("-e", ".")
     session.install("-e", ".[pandas,storage]")
 
     # Run py.test against the unit tests.
@@ -117,7 +116,6 @@ def system(session):
     for local_dep in LOCAL_DEPS:
         session.install("-e", local_dep)
     session.install("-e", "../test_utils/")
-    session.install("-e", ".")
     session.install("-e", ".[pandas,storage]")
 
     # Run py.test against the system tests.
@@ -144,7 +142,7 @@ def cover(session):
 def docs(session):
     """Build the docs for this library."""
 
-    session.install("-e", ".")
+    session.install("-e", ".[pandas,storage]")
     session.install("sphinx", "alabaster", "recommonmark")
 
     shutil.rmtree(os.path.join("docs", "_build"), ignore_errors=True)