+++ storage/google/cloud/storage/client.py
@@ -22,6 +22,7 @@ from google.api_core import page_iterator
 from google.cloud._helpers import _LocalStack
 from google.cloud.client import ClientWithProject
 from google.cloud.exceptions import NotFound
+from google.cloud.storage._helpers import _get_storage_host
 from google.cloud.storage._http import Connection
 from google.cloud.storage.batch import Batch
 from google.cloud.storage.bucket import Bucket
@@ -94,6 +95,9 @@ class Client(ClientWithProject):
         )
 
         kw_args = {"client_info": client_info}
+
+        kw_args["api_endpoint"] = _get_storage_host()
+
         if client_options:
             if type(client_options) == dict:
                 client_options = google.api_core.client_options.from_dict(
@@ -578,7 +582,9 @@ class Client(ClientWithProject):
             extra_params=extra_params,
         )
 
-    def create_hmac_key(self, service_account_email, project_id=None):
+    def create_hmac_key(
+        self, service_account_email, project_id=None, user_project=None
+    ):
         """Create an HMAC key for a service account.
 
         :type service_account_email: str
@@ -587,6 +593,9 @@ class Client(ClientWithProject):
         :type project_id: str
         :param project_id: (Optional) explicit project ID for the key.
             Defaults to the client's project.
+
+        :type user_project: str
+        :param user_project: (Optional) This parameter is currently ignored.
 
         :rtype:
             Tuple[:class:`~google.cloud.storage.hmac_key.HMACKeyMetadata`, str]
@@ -597,6 +606,10 @@ class Client(ClientWithProject):
 
         path = "/projects/{}/hmacKeys".format(project_id)
         qs_params = {"serviceAccountEmail": service_account_email}
+
+        if user_project is not None:
+            qs_params["userProject"] = user_project
+
         api_response = self._connection.api_request(
             method="POST", path=path, query_params=qs_params
         )
@@ -611,6 +624,7 @@ class Client(ClientWithProject):
         service_account_email=None,
         show_deleted_keys=None,
         project_id=None,
+        user_project=None,
     ):
         """List HMAC keys for a project.
 
@@ -631,6 +645,9 @@ class Client(ClientWithProject):
         :param project_id: (Optional) explicit project ID for the key.
             Defaults to the client's project.
 
+        :type user_project: str
+        :param user_project: (Optional) This parameter is currently ignored.
+
         :rtype:
             Tuple[:class:`~google.cloud.storage.hmac_key.HMACKeyMetadata`, str]
         :returns: metadata for the created key, plus the bytes of the key's secret, which is an 40-character base64-encoded string.
@@ -647,6 +664,9 @@ class Client(ClientWithProject):
         if show_deleted_keys is not None:
             extra_params["showDeletedKeys"] = show_deleted_keys
 
+        if user_project is not None:
+            extra_params["userProject"] = user_project
+
         return page_iterator.HTTPIterator(
             client=self,
             api_request=self._connection.api_request,
@@ -656,7 +676,7 @@ class Client(ClientWithProject):
             extra_params=extra_params,
         )
 
-    def get_hmac_key_metadata(self, access_id, project_id=None):
+    def get_hmac_key_metadata(self, access_id, project_id=None, user_project=None):
         """Return a metadata instance for the given HMAC key.
 
         :type access_id: str
@@ -665,8 +685,11 @@ class Client(ClientWithProject):
         :type project_id: str
         :param project_id: (Optional) project ID of an existing key.
             Defaults to client's project.
+
+        :type user_project: str
+        :param user_project: (Optional) This parameter is currently ignored.
         """
-        metadata = HMACKeyMetadata(self, access_id, project_id)
+        metadata = HMACKeyMetadata(self, access_id, project_id, user_project)
         metadata.reload()  # raises NotFound for missing key
         return metadata