+++ tasks/google/cloud/tasks_v2/gapic/cloud_tasks_client.py
@@ -1223,7 +1223,7 @@ class CloudTasksClient(object):
         Args:
             parent (str): Required. The queue name. For example:
                 ``projects/PROJECT_ID/locations/LOCATION_ID/queues/QUEUE_ID``
-            response_view (~google.cloud.tasks_v2.types.View): The response\_view specifies which subset of the ``Task`` will be
+            response_view (~google.cloud.tasks_v2.enums.Task.View): The response\_view specifies which subset of the ``Task`` will be
                 returned.
 
                 By default response\_view is ``BASIC``; not all information is retrieved
@@ -1326,7 +1326,7 @@ class CloudTasksClient(object):
         Args:
             name (str): Required. The task name. For example:
                 ``projects/PROJECT_ID/locations/LOCATION_ID/queues/QUEUE_ID/tasks/TASK_ID``
-            response_view (~google.cloud.tasks_v2.types.View): The response\_view specifies which subset of the ``Task`` will be
+            response_view (~google.cloud.tasks_v2.enums.Task.View): The response\_view specifies which subset of the ``Task`` will be
                 returned.
 
                 By default response\_view is ``BASIC``; not all information is retrieved
@@ -1451,7 +1451,7 @@ class CloudTasksClient(object):
 
                 If a dict is provided, it must be of the same form as the protobuf
                 message :class:`~google.cloud.tasks_v2.types.Task`
-            response_view (~google.cloud.tasks_v2.types.View): The response\_view specifies which subset of the ``Task`` will be
+            response_view (~google.cloud.tasks_v2.enums.Task.View): The response\_view specifies which subset of the ``Task`` will be
                 returned.
 
                 By default response\_view is ``BASIC``; not all information is retrieved
@@ -1626,7 +1626,7 @@ class CloudTasksClient(object):
         Args:
             name (str): Required. The task name. For example:
                 ``projects/PROJECT_ID/locations/LOCATION_ID/queues/QUEUE_ID/tasks/TASK_ID``
-            response_view (~google.cloud.tasks_v2.types.View): The response\_view specifies which subset of the ``Task`` will be
+            response_view (~google.cloud.tasks_v2.enums.Task.View): The response\_view specifies which subset of the ``Task`` will be
                 returned.
 
                 By default response\_view is ``BASIC``; not all information is retrieved